+++ CHANGELOG.md
@@ -835,6 +835,7 @@ Promptbook server has OpenAI compatibility
 -   Alongside `OpenAiExecutionTools` and `OpenAiAssistantExecutionTools` add `OpenAiCompatibleExecutionTools` as registration of configuration and the constructor
 -   Logging all failed results not just last result
 -   Do not cache failed results and bring `DEFAULT_MAX_EXECUTION_ATTEMPTS` down to 7
+-   Make `gpt-4-turbo` default "vanilla" chat model of `OpenAiExecutionTools`
 
 ## Upcoming